no leaks, still have to fix player on labrinth map

+++ libft/alloc_matrix.c
@@ -6,7 +6,7 @@
 /*   By: tmadau <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/09/11 08:12:13 by tmadau            #+#    #+#             */
-/*   Updated: 2018/09/12 14:14:44 by tmadau           ###   ########.fr       */
+/*   Updated: 2018/09/17 12:06:07 by tmadau           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,10 +17,10 @@ void	alloc_matrix(char *str, t_map *ev)
 	int		index;
 
 	index = 0;
-	ev->map = (double **)malloc(sizeof(double *) * ev->y);
+	ev->map = (int **)malloc(sizeof(int *) * ev->y);
 	while (index < ev->y)
 	{
-		ev->map[index] = (double *)malloc(sizeof(double) * ev->x);
+		ev->map[index] = (int *)malloc(sizeof(int) * ev->x);
 		index++;
 	}
 	fill_matrix(str, ev);
